Typo Fixed: MIME Content Type is image/ not images/
Per RFC 2045, Section 5.1, Content Type is "image" not "images". https://tools.ietf.org/html/rfc2045

Additionally, the ext portion is user-controlled and may not be in lowercase as expected in the cType dictionary.  Added .lower() when called to allow "JPG" and "jpg" to both match the "jpg" lookup for "image/jpeg".