Added further information for newbie users
index.GET is a bit confused for non-programmers, the 'index' name is repeated through out the project, and is not apparent the dot is used as concatenation as opposed to a period. Mentioning the method GET, makes it much easier for a new user to understand where the modification is taking place.

Added note on MySQL Create Table statement.